BP_RedFlag Blueprint was causing crashes, fixed by deleting and remaking BP_RedFlag from a child of BP_Flag. Did the same to BP_BlueFlag just in case.
Adjusted health values for BP_BlasterCharacter for 100 health and 150 shield.

Reduced size of pickup radius for pickup buffs to match their visual size.

Reduced size of Grenade Launcher and Rocket Launcher ammo to not be as excessivly large.

Created sub folders in the Maps folder to organize between GameplayMaps, GameSystemMaps, and TestMaps